Refactor (a): the client's answer to a key is a receptacle, not a step file
Per architecture-first-principles' C3 — the graduation sweep. The fix
above left ~140 lines of machinery inside `editing_steps.ts` whose whole
subject is one hard, volatile thing: HOW THE HARNESS KNOWS THE CLIENT HAS
ANSWERED A KEY. That is not the row editor's steps; it is a ritual, and
the file next to it exists for exactly this reason — `support/said.ts`
was pulled out because "two features asking the same question two
different ways is how one of them stops being asked properly", and three
step files could want this one.

So `support/caret.ts`, named for the noun the whole thing is about. The
answer has three shapes and now says so once: the caret LEAVES the line a
key ended, the row it is in is drawn where a key MOVED it, or it comes
BACK to the line — and every one of them will take "the page said why it
did not" instead.

Two duplicates went with the move (C2, consumer ergonomics):

- the click-away wait and the move-key wait were the same predicate
  spelled twice; both are `leftThePlace` now;
- `editorHeld(world, which)` had one caller and one value of `which`.
  It is `lineHeld(world)`.

No behaviour change. The full suite, twice, under nproc-saturating load:
green both times.
